requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.51
pytest==8.2.0
requests==2.31.0
selenium==4.20.0
